+++ finagle-thriftmux/src/test/scala/com/twitter/finagle/thriftmux/ThriftSmuxTest.scala
@@ -150,7 +150,7 @@ class ThriftSmuxTest extends FunSuite {
             results.get
           }
           assert(string.isEmpty)
-          assert(stats.counters.get(Seq("client", "failures")) == None)
+          assert(stats.counters.get(Seq("client", "failures")) == Some(0))
           assert(stats.counters.get(Seq("client", "service_creation", "failures")) == Some(1))
           assert(
             stats.counters.get(